% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_hists.R
\name{age_hist}
\alias{age_hist}
\title{Escapement by Age Histogram}
\usage{
age_hist(samples, file_path, width = 500, height = 1000, final_year = 2020)
}
\arguments{
\item{samples}{(data.frame) output of clean_chains}

\item{file_path}{(character) file path to store output jpeg file.}

\item{width}{(numeric) width of the output plot (in px).}

\item{height}{(numeric) height of the output plot (in px).}
}
\description{
Escapement by Age Histogram
}
