% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_chains.R
\name{clean_chains}
\alias{clean_chains}
\title{Convert chains to an easier to use format and burnin further if the burnin period was not long enough}
\usage{
clean_chains(samples, burnin_further = 0)
}
\arguments{
\item{samples}{Output of jagsUI::jags}

\item{burnin_further}{The number of additional MCMC iterations to discard as burnin}
}
\description{
Convert chains to an easier to use format and burnin further if the burnin period was not long enough
}
