% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_abundance_estimates.R
\name{extract_abundance_estimates}
\alias{extract_abundance_estimates}
\title{Extract abundance estimates from Jags output}
\usage{
extract_abundance_estimates(samples, alpha = 0.1, final_year = 2030, file_path)
}
\arguments{
\item{samples}{(data.frame) output of clean_chains}

\item{alpha}{(numeric) desired significance level for credible intervals}

\item{file_path}{(character) file path to store output xlsx file.}
}
\description{
Extract abundance estimates from Jags output
}
