% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_age_at_maturity.R
\name{extract_age_at_maturity}
\alias{extract_age_at_maturity}
\title{Extract estimates of total run abundance by age from Jags output}
\usage{
extract_age_at_maturity(samples, alpha, file_path, final_year = 2020)
}
\arguments{
\item{samples}{(data.frame) output of clean_chains}

\item{alpha}{(numeric) desired significance level for credible intervals}

\item{file_path}{(character) file path to store output xlsx file.}
}
\description{
Extract estimates of total run abundance by age from Jags output
}
