% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sr_params.R
\name{extract_sr_params}
\alias{extract_sr_params}
\title{Extract biological reference points and parameters contained in the SR relationship}
\usage{
extract_sr_params(samples, model, alpha, file_path)
}
\arguments{
\item{samples}{(data.frame) output of clean_chains}

\item{alpha}{(numeric) desired significance level for credible intervals}

\item{file_path}{(character) file path to store output xlsx file}
}
\description{
Extract biological reference points and parameters contained in the SR relationship
}
