% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.standalone}
\alias{format.standalone}
\title{Format the data to run a paired analysis}
\usage{
\method{format}{standalone}(data)
}
\description{
The output of the function will be used as data within the Jags model. Function is used internally within get_jags_data
}
