% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_inits.R
\name{get_new_inits}
\alias{get_new_inits}
\title{Extracts and formats end values of chains so that they can be used as initial values in new run}
\usage{
get_new_inits(jags_out, n_chains)
}
\arguments{
\item{jags_out}{Output of jagsUI::jags}

\item{n_chains}{Number of chains in MCMC routine}
}
\description{
Extracts and formats end values of chains so that they can be used as initial values in new run
}
