% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Retrieves the appropriate parameters to monitor}
\usage{
get_params(model)
}
\arguments{
\item{model}{(character) the model that is being run. Accepted values are "base", "base_tvm", "base_ar", "base_tvm_ar", and "base_tvp".}
}
\description{
Retrieves the appropriate parameters to monitor
}
