% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horsetail_plot.R
\name{horsetail_plot}
\alias{horsetail_plot}
\title{Horsetail Plot}
\usage{
horsetail_plot(
  samples,
  file_path,
  n_draws = 25,
  sig_lev = 0.5,
  width = 800,
  height = 800,
  s_up = 20000,
  r_up = c(30000, 30000)
)
}
\arguments{
\item{samples}{(data.frame) output of clean_chains}

\item{file_path}{(character) file path to store output xlsx file}

\item{n_draws}{(numeric) the number of SR relations to be simulated}

\item{sig_lev}{(numeric) significance level for the error bars}

\item{width}{(numeric) width of the output plot (in px)}

\item{height}{(numeric) height of the output plot (in px)}

\item{s_up}{(numeric) vector of length 1 specifying the upper bounds for the x axis.}

\item{r_up}{(numeric) vector of length 2 specifying the upper bound for the y axis. The first entry cooresponds to the Chena while the second entry cooresponds to the Salcha.}
}
\description{
Horsetail Plot
}
