% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{profile_plot}
\alias{profile_plot}
\title{Optimal Yield, Overfishing, and Optimal Recruitment Probability Plot}
\usage{
profile_plot(
  samples,
  model,
  file_path,
  old_goal_med,
  old_goal_lo,
  old_goal_up,
  legend_position = c(15000, 1),
  width = 1000,
  height = 1000
)
}
\arguments{
\item{samples}{(data.frame) output of clean_chains.}

\item{file_path}{(character) file path to store the output jpeg file.}

\item{old_goal_lo}{(numeric) vector of length 2 specifying the lower bound of the old escapement goal. The first entry cooresponds to the Chena while the second entry cooresponds to the Salcha.}

\item{old_goal_up}{(numeric)  vector of length 2 specifying the upper bound of the old escapement goal. The first entry cooresponds to the Chena while the second entry cooresponds to the Salcha.}

\item{legend_position}{(numeric) vector of length 2 specifying the x and y coordinate where the legend should be positioned in each plot.}

\item{width}{(numeric) the width of the output jpeg file (in px).}

\item{height}{(numeric) the height of the output jpeg file (in px).}

\item{old_goal_median}{(numeric) vector of length 2 specifying the old escapement goal. The first entry cooresponds to the Chena while the second entry cooresponds to the Salcha.}
}
\description{
Optimal Yield, Overfishing, and Optimal Recruitment Probability Plot
}
