% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residual_plot}
\alias{residual_plot}
\title{Plot residuals of a model -- need to implement for all models, but should work for base model}
\usage{
residual_plot(samples, file_path, width = 750, height = 500)
}
\arguments{
\item{samples}{(data.frame) output of clean_chains.}

\item{file_path}{(character) file path to store the output jpeg file.}

\item{width}{(numeric) the width of the output jpeg file (in px).}

\item{height}{(numeric) the height of the output jpeg file (in px).}
}
\description{
Plot residuals of a model -- need to implement for all models, but should work for base model
}
