% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_jags_model.R
\name{save_jags_model}
\alias{save_jags_model}
\title{Save .jags version of a JAGS module
Modified JAGS modules need saved for the changes to take effect. This function streamlines the process. Jags modules are base.R, base_ar.R, base_tvm.R, base_tvm_ar.R, base_tvp.R, base_wl.R, and bev_holt.R}
\usage{
save_jags_model(mod_path, model)
}
\arguments{
\item{mod_path}{(character) path to save the .jags file}

\item{model}{(character) the model to save. accepted values are "base", "base_tvm", "base_ar", "base_tvm_ar", "base_tvp", "base_wl", and "bev_holt"}
}
\description{
Save .jags version of a JAGS module
Modified JAGS modules need saved for the changes to take effect. This function streamlines the process. Jags modules are base.R, base_ar.R, base_tvm.R, base_tvm_ar.R, base_tvp.R, base_wl.R, and bev_holt.R
}
