% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_run_info.R
\name{save_run_info}
\alias{save_run_info}
\title{Save useful model run info #}
\usage{
save_run_info(
  model_specs,
  run_time,
  DIC,
  pD,
  file_path = "Tables/run_info.xlsx",
  exists = T
)
}
\arguments{
\item{model_specs}{(list) Named list specifying the run specifications.}

\item{run_time}{(difftime) The time it took for the model to run.}

\item{DIC}{(numeric) DIC reported by JAGS.}

\item{file_path}{(character) File path to save model run information.}

\item{exists}{(logical) If a model run information table already exists, use exists=T. Use exists=F otherwise.}
}
\description{
Save useful model run info #
}
