% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_directory.R
\name{source_directory}
\alias{source_directory}
\title{Source all files in a directory}
\usage{
source_directory(directory = "S:/ChenaSalchaSR2020/Rfunctions")
}
\arguments{
\item{directory}{(character) the directory in which all files will be source}
}
\description{
Source all files in a directory
}
