% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sustainable_yield.R
\name{sy_plot}
\alias{sy_plot}
\title{Sustainable Yield Plot}
\usage{
sy_plot(
  samples,
  model,
  file_path,
  alpha = 0.5,
  leg_pos = c(17500, 40000),
  width = 750,
  height = 750,
  y_up = 70000
)
}
\arguments{
\item{samples}{(data.frame) output of clean_chains.}

\item{file_path}{(character) file path to store the output jpeg file.}

\item{alpha}{(numeric) desired significance level for credible intervals.}

\item{leg_pos}{(numeric) vector of length 2 specifing the x and y coordinate where the legend should be positioned.}

\item{width}{(numeric) width of the output plot (in px).}

\item{height}{(numeric) height of the output plot (in px).}
}
\description{
Sustainable Yield Plot
}
