% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_params.R
\name{trace_params}
\alias{trace_params}
\title{Get desired parameters for traceplots}
\usage{
trace_params(params)
}
\arguments{
\item{params}{(character vector) output of get_params}
}
\description{
Get desired parameters for traceplots
}
