% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ar.R
\name{write_jags_model.base_ar}
\alias{write_jags_model.base_ar}
\title{AR(1) term added to base Ricker}
\usage{
write_jags_model.base_ar(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
AR(1) term added to base Ricker
}
