% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_ld.R
\name{write_jags_model.base_ld}
\alias{write_jags_model.base_ld}
\title{basic Ricker model}
\usage{
write_jags_model.base_ld(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
basic Ricker model
}
