% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tvm.R
\name{write_jags_model.base_tvm}
\alias{write_jags_model.base_tvm}
\title{Time varying age-at-maturity added to base Ricker}
\usage{
write_jags_model.base_tvm(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
Time varying age-at-maturity added to base Ricker
}
