% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tvm_ar_ash_2.R
\name{write_jags_model.base_tvm_ar_ash}
\alias{write_jags_model.base_tvm_ar_ash}
\title{Age-Stratified Harvest Model}
\usage{
write_jags_model.base_tvm_ar_ash(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
Age-Stratified Harvest Model
}
