% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tvm_ar_tvp.R
\name{write_jags_model.base_tvm_ar_tvp}
\alias{write_jags_model.base_tvm_ar_tvp}
\title{Time varying age-at-maturity and AR(1) term added to base Ricker}
\usage{
write_jags_model.base_tvm_ar_tvp(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
Time varying age-at-maturity and AR(1) term added to base Ricker
}
