% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tvm_ar_ash.R, R/temp.R
\name{write_jags_model.base_tvm_ash}
\alias{write_jags_model.base_tvm_ash}
\title{time varying age-at-maturity, AR(1) term, and age-stratified harvest terms added to base model}
\usage{
write_jags_model.base_tvm_ash(path)

write_jags_model.base_tvm_ash(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
time varying age-at-maturity, AR(1) term, and age-stratified harvest terms added to base model

time varying age-at-maturity, AR(1) term, and age-stratified harvest terms added to base model
}
