% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tvm_tvp.R
\name{write_jags_model.base_tvm_tvp}
\alias{write_jags_model.base_tvm_tvp}
\title{Base Ricker with time varying productivity}
\usage{
write_jags_model.base_tvm_tvp(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
Base Ricker with time varying productivity
}
