% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_tvp_ld.R
\name{write_jags_model.base_tvp_ld}
\alias{write_jags_model.base_tvp_ld}
\title{Base Ricker with time varying productivity}
\usage{
write_jags_model.base_tvp_ld(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
Base Ricker with time varying productivity
}
