% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_wl.R
\name{write_jags_model.base_wl}
\alias{write_jags_model.base_wl}
\title{Base Ricker with a term for the mean August water level}
\usage{
write_jags_model.base_wl(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
Base Ricker with a term for the mean August water level
}
