% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bev_holt.R
\name{write_jags_model.bev_holt}
\alias{write_jags_model.bev_holt}
\title{Basic Beverton-Holt model}
\usage{
write_jags_model.bev_holt(path)
}
\arguments{
\item{path}{(character) path to write .jags file}
}
\description{
Basic Beverton-Holt model
}
