% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_inits.R, R/get_initial_values.R
\name{get_initial_values}
\alias{get_initial_values}
\title{Get a set of initial values for alpha and beta}
\usage{
get_initial_values()

get_initial_values()
}
\description{
Get a set of initial values for alpha and beta

Get a set of initial values for alpha and beta
}
