% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jags_data.R
\name{get_jags_data}
\alias{get_jags_data}
\title{Read in and format data for JAGS}
\usage{
get_jags_data(file_path, model)
}
\arguments{
\item{file_path}{(String) Directory to the xlsx data. Note: the format of the excel table needs to be standardized (see Data/2020-Data.xlsx).}

\item{type}{(String) The type of analysis to run. Accepted values are "Paired", "Chena", and "Salcha".}
}
\description{
Read in and format data for JAGS
}
