% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_model_specifications.R
\name{set_model_specifications}
\alias{set_model_specifications}
\title{Specify important information related to model run}
\usage{
set_model_specifications(
  model_name,
  notes,
  n_chains = 4,
  n_iter = 5e+05,
  n_burnin = 250000,
  n_thin = 100,
  parallel = F
)
}
\arguments{
\item{notes}{(character) store a note related to the model run to keep track of why the model is being run with the chosen specs}

\item{n_chains}{(numeric) the number of chains}

\item{n_iter}{(numeric) the number of iterations including burnin}

\item{n_burnin}{(numeric) the number of iterations to discard as burnin}

\item{n_thin}{(numeric) the thinning period}

\item{parallel}{(logical) use parallel = F, or find a working wrapper for parallel computation. jagsUI::jags is a little buggy when parallel=T}

\item{run_name}{(character) run identifier for output tables and plots. keep it short and informative}
}
\description{
Specify important information related to model run
}
